تقرير بالعمليات للموظف يكون بفترات زمنية مختلفة يتم تحديدها

+++ Modules/CourseModule/app/Http/Controllers/Admin/CourseAdminController.php
@@ -12,6 +12,8 @@ use Modules\CourseModule\Services\CourseRegPaymentService;
 use Modules\CourseModule\Services\CourseRegService;
 use Modules\CourseModule\Services\CourseRegStatusService;
 use Modules\CourseModule\Services\CourseService;
+use Modules\LogModule\app\Http\Models\Log;
+use Modules\LogModule\Services\LogService;
 
 class CourseAdminController extends Controller
 {
@@ -20,14 +22,16 @@ class CourseAdminController extends Controller
     private $courseRegService;
     private $courseRegPaymentService;
     private $courseRegStatusService;
+    private $logService;
 
-    public function __construct(CourseService $courseService, BranchService $branchService, CourseRegService $courseRegService, CourseRegPaymentService $courseRegPaymentService, CourseRegStatusService $courseRegStatusService)
+    public function __construct(CourseService $courseService, BranchService $branchService, CourseRegService $courseRegService, CourseRegPaymentService $courseRegPaymentService, CourseRegStatusService $courseRegStatusService, LogService $logService)
     {
         $this->courseService = $courseService;
         $this->branchService = $branchService;
         $this->courseRegService = $courseRegService;
         $this->courseRegPaymentService = $courseRegPaymentService;
         $this->courseRegStatusService = $courseRegStatusService;
+        $this->logService = $logService;
     }
 
     /**
@@ -97,7 +101,11 @@ class CourseAdminController extends Controller
                 ->withInput();
         }
         $this->courseService->create($request);
-
+        //Add Log
+        $action = 'إضافة';
+        $description = 'تم إضافة دورة : ' . $request->name;
+        $this->logService->recordLog($action, $description, url()->current());
+        //
         return redirect()->route(Auth::getDefaultDriver() . '.courses')
             ->with('success', 'تم الاضافه بنجاح.');
     }
@@ -180,6 +188,11 @@ class CourseAdminController extends Controller
         }
         $this->courseService->update($request);
 
+        //Add Log
+        $action = 'تعديل';
+        $description = 'تم تعديل دورة : ' . $request->name;
+        $this->logService->recordLog($action, $description, url()->current());
+        //
         return redirect()->route(Auth::getDefaultDriver() . '.courses.show', $request->id)
             ->with('success', 'تم التعديل بنجاح.');
     }
@@ -191,7 +204,14 @@ class CourseAdminController extends Controller
      */
     public function destroy($id)
     {
+        $course = $this->courseService->findOne($id);
         $this->courseService->deleteOne($id);
+
+        //Add Log
+        $action = 'حذف';
+        $description = 'تم حذف دورة : ' . $course->name;
+        $this->logService->recordLog($action, $description, url()->current());
+        //
         return redirect()->route(Auth::getDefaultDriver() . '.courses')
             ->with('success', 'حذف الدورة بنجاح.');
     }
@@ -212,8 +232,13 @@ class CourseAdminController extends Controller
                 ->withErrors('عفوا... تم التسجيل في هذه الدورة من قبل')
                 ->withInput();
         // dd($request->all());
-        $this->courseRegService->registerStudentToCourse($request);
+        $course_reg = $this->courseRegService->registerStudentToCourse($request);
 
+        //Add Log
+        $action = 'تسجيل في دورة';
+        $description = 'تم تسجيل طالب في دورة : ' . $course_reg->course->name . ' - ' . $course_reg->student->name;
+        $this->logService->recordLog($action, $description, url()->current());
+        //
         return redirect()->route(Auth::getDefaultDriver() . '.students.view', $request->student_id)
             ->with('success', 'تم الاضافه بنجاح ...');
     }
@@ -233,6 +258,19 @@ class CourseAdminController extends Controller
         }
         $courseDate = $this->courseRegService->takeRegAction($request->all());
 
+        //Add Log
+        if ($request->action == 'accept') {
+            $action = 'قبول';
+            $description = 'تم قبول تسجيل طالب في دورة : ' . $courseDate->course->name . ' - ' . $courseDate->student->name;
+        } elseif ($request->action == 'reject') {
+            $action = 'رفض';
+            $description = 'تم رفض تسجيل طالب في دورة : ' . $courseDate->course->name . ' - ' . $courseDate->student->name;
+        } elseif ($request->action == 'waiting') {
+            $action = 'إنتظار';
+            $description = 'تم وضع تسجيل طالب في دورة في حالة إنتظار : ' . $courseDate->course->name . ' - ' . $courseDate->student->name;
+        }
+        $this->logService->recordLog($action, $description, url()->current());
+        //
         return back()
             ->with('success', 'تمت العمليه بنجاح.');
     }
@@ -263,10 +301,15 @@ class CourseAdminController extends Controller
                 ->withErrors($validator)
                 ->withInput();
         }
-        $this->courseRegPaymentService->takePayAction($request->all());
+        $course_reg = $this->courseRegPaymentService->takePayAction($request->all());
 
         $this->courseRegService->checkAndUpdateRegStatus($request->id);
 
+        //Add Log
+        $action = 'دفع';
+        $description = 'تم تسجيل دفعة مالية لدورة : ' . $course_reg->course->name . ' - مقابل : ' . $request->pay_type . ' - ' . $course_reg->student->name . ' - المبلغ : ' . $request->amount;
+        $this->logService->recordLog($action, $description, url()->current());
+        //
         return back()
             ->with('success', 'تمت العمليه بنجاح.');
     }
@@ -286,7 +329,22 @@ class CourseAdminController extends Controller
                 ->withErrors($validator)
                 ->withInput();
         }
-        $this->courseRegService->updatePaymentType($request);
+        $course_reg = $this->courseRegService->updatePaymentType($request);
+
+        //Add Log
+        if ($request->status == 'normal') {
+            $status_text = 'افتراضي';
+        } elseif ($request->status == 'free') {
+            $status_text = 'الدورة مجانا';
+        } elseif ($request->status == 'nopaying') {
+            $status_text = 'ممتنع عن السداد';
+        } elseif ($request->status == 'leave') {
+            $status_text = 'مغادر';
+        }
+        $action = 'تعديل طريقة الدفع';
+        $description = 'تم تعديل طريقة الدفع في دورة : ' . $course_reg->course->name . ' - ' . $course_reg->student->name . ' - الي طريقة الدفع : ' . $status_text;
+        $this->logService->recordLog($action, $description, url()->current());
+        //
         return back()
             ->with('success', 'تمت العمليه بنجاح.');
     }
@@ -306,7 +364,12 @@ class CourseAdminController extends Controller
                 ->withErrors($validator)
                 ->withInput();
         }
-        $this->courseRegService->updateRegStatus($request);
+        $course_reg = $this->courseRegService->updateRegStatus($request);
+        //Add Log
+        $action = 'تعديل حالة التسجيل';
+        $description = 'تم تعديل حالة التسجيل في دورة : ' . $course_reg->course->name . ' - ' . $course_reg->student->name . ' - الي حالة : ' . $course_reg->status->status;
+        $this->logService->recordLog($action, $description, url()->current());
+        //
         return back()
             ->with('success', 'تمت العمليه بنجاح.');
     }
@@ -339,15 +402,30 @@ class CourseAdminController extends Controller
 
     public function destroyReg($id)
     {
+        $course_reg = $this->courseRegService->findOne($id);
+        if (Auth::guard('user')->user()->branch_id != $course_reg->course->branch_id)
+            return back()
+                ->withErrors('الدورة غير موجود في قائمة الدورات');
         $this->courseRegService->deleteOne($id);
+        //Add Log
+        $action = 'حذف تسجيل في دورة';
+        $description = 'تم حذف تسجيل في دورة : ' . $course_reg->course->name . ' - ' . $course_reg->student->name;
+        $this->logService->recordLog($action, $description, url()->current());
+        //
         return back()
             ->with('success', 'تم حذف الموعد بنجاح.');
     }
 
     public function setCertDelivered($id)
     {
-        $this->courseRegService->updateCertDelivered($id, 1);
+        $course_reg = $this->courseRegService->updateCertDelivered($id, 1);
         $this->courseRegService->checkAndUpdateRegStatus($id);
+
+        //Add Log
+        $action = 'استلام شهاده';
+        $description = 'تم تسجيل استلام الشهاده في دورة : ' . $course_reg->course->name . ' - ' . $course_reg->student->name;
+        $this->logService->recordLog($action, $description, url()->current());
+        //
 
         return back()
             ->with('success', 'تم استلام الشهاده بنجاح.');
@@ -355,50 +433,73 @@ class CourseAdminController extends Controller
 
     public function setCertNotDelivered($id)
     {
-        $this->courseRegService->updateCertDelivered($id, 0);
+        $course_reg = $this->courseRegService->updateCertDelivered($id, 0);
         $this->courseRegService->checkAndUpdateRegStatus($id);
-
+        //Add Log
+        $action = 'استلام شهاده';
+        $description = 'تم تسجيل عدم استلام الشهاده في دورة : ' . $course_reg->course->name . ' - ' . $course_reg->student->name;
+        $this->logService->recordLog($action, $description, url()->current());
+        //
         return back()
             ->with('success', 'تم تعديل عدم استلام الشهاده بنجاح.');
     }
 
     public function ChangePriceForOneStudent(Request $request)
     {
-        $price = $this->courseRegService->updatePriceForOneStudent($request);
+        $course_reg = $this->courseRegService->updatePriceForOneStudent($request);
         $this->courseRegService->checkAndUpdateRegStatus($request->reg_id);
+
+        //Add Log
+        $action = 'تعديل سعر';
+        $description = 'تم تعديل السعر في دورة : ' . $course_reg->course->name . ' - ' . $course_reg->student->name . ' - الي سعر :' .  $course_reg->student_price;
+        $this->logService->recordLog($action, $description, url()->current());
+        //
         return response()->json(
             array(
                 'success' => "true",
-                'new_price' => $price->price,
-                'id' => $price->id,
+                'new_price' => $course_reg->student_price,
+                'id' => $course_reg->id,
 
             )
         );
     }
 
-     public function UpdateDiscountForOneStudent(Request $request)
+    public function UpdateDiscountForOneStudent(Request $request)
     {
-        $price = $this->courseRegService->updateDiscountForOneStudent($request);
+        $course_reg = $this->courseRegService->updateDiscountForOneStudent($request);
         $this->courseRegService->checkAndUpdateRegStatus($request->reg_id);
+
+        //Add Log
+        $action = 'تعديل خصم';
+        $description = 'تم تعديل خصم دورة : ' . $course_reg->course->name . ' - ' . $course_reg->student->name . ' - الي خصم :' .  $course_reg->discount_amount;
+        $this->logService->recordLog($action, $description, url()->current());
+        //
+
         return response()->json(
             array(
                 'success' => "true",
-                'new_price' => $price->price,
-                'new_discount' => $price->discount_amount,
-                'id' => $price->id,
+                'new_price' => $course_reg->price,
+                'new_discount' => $course_reg->discount_amount,
+                'id' => $course_reg->id,
             )
         );
     }
 
     public function ChangeExamPriceForOneStudent(Request $request)
     {
-        $price = $this->courseRegService->updateExamPriceForOneStudent($request);
+        $course_reg = $this->courseRegService->updateExamPriceForOneStudent($request);
         $this->courseRegService->checkAndUpdateRegStatus($request->reg_id);
+
+        //Add Log
+        $action = 'تعديل رسوم اختبار';
+        $description = 'تم تعديل رسوم الاختبار في دورة : ' . $course_reg->course->name . ' - ' . $course_reg->student->name . ' - الي رسوم اختبار :' .  $course_reg->exam_fees;
+        $this->logService->recordLog($action, $description, url()->current());
+        //
         return response()->json(
             array(
                 'success' => "true",
-                'new_exam_price' => $price->exam_fees,
-                'id' => $price->id,
+                'new_exam_price' => $course_reg->exam_fees,
+                'id' => $course_reg->id,
 
             )
         );
