+++ Modules/StudentModule/app/Http/Controllers/Admin/StudentAdminController.php
@@ -15,6 +15,8 @@ use Modules\StudentModule\Exports\StudentExport;
 use Maatwebsite\Excel\Facades\Excel;
 use Modules\CourseModule\Services\CourseService;
 use Illuminate\Validation\Rule;
+use Modules\LogModule\app\Http\Models\Log;
+use Modules\LogModule\Services\LogService;
 
 class StudentAdminController extends Controller
 {
@@ -22,13 +24,15 @@ class StudentAdminController extends Controller
     private $branchService;
     private $cityService;
     private $courseService;
+    private $logService;
 
-    public function __construct(StudentService $studentService, BranchService $branchService, CityService $cityService, CourseService $courseService)
+    public function __construct(StudentService $studentService, BranchService $branchService, CityService $cityService, CourseService $courseService,LogService $logService)
     {
         $this->studentService = $studentService;
         $this->branchService = $branchService;
         $this->cityService = $cityService;
         $this->courseService = $courseService;
+        $this->logService = $logService;
     }
 
     public function add()
@@ -79,6 +83,11 @@ class StudentAdminController extends Controller
 
         $student = $this->studentService->create($request);
 
+        //Add Log
+        $action = 'إضافة طالب';
+        $description = 'تم إضافة الطالب ' . $student->name . ' برقم هوية ' . $student->id_nu;
+        $this->logService->recordLog($action, $description, url()->current());
+        //
         return redirect()->route(Auth::getDefaultDriver() . '.students.view', $student->id)
             ->with('success', 'تم الاضافه بنجاح ...');
     }
@@ -158,18 +167,31 @@ class StudentAdminController extends Controller
                 ->withErrors($validator)
                 ->withInput();
         }
-        $this->studentService->update($request);
+        $student = $this->studentService->update($request);
 
+        //Add Log
+        $action = 'تعديل بيانات طالب';
+        $description = 'تم تعديل بيانات الطالب ' . $student->name ;
+        $this->logService->recordLog($action, $description, url()->current());
+        //
         return redirect()->route(Auth::getDefaultDriver() . '.students.view', $request->id)
             ->with('success', 'Updated Successfully.');
     }
 
     public function destroy($id)
     {
-        // if (!in_array('can_del_students', auth('admin')->user()->privileges_keys()))
-        //     return back()->withErrors('لا يمكن اتمام هذه العمليه');
+        $student = $this->studentService->findOne($id);
+        if (!$student)
+            return back()
+                ->withErrors('الطالب غير موجود في قائمة الطلاب');   
 
         $this->studentService->deleteOne($id);
+
+        //Add Log
+        $action = 'حذف طالب';
+        $description = 'تم حذف الطالب ' . $student->name . ' برقم هوية ' . $student->id_nu;
+        $this->logService->recordLog($action, $description, url()->current());
+        //
         return redirect()->route(Auth::getDefaultDriver() . '.students')
             ->with('success', 'Deleted Successfully.');
     }
